+++ Changelog.md
@@ -40,6 +40,7 @@
 * Modify Name Files Project (19/11/2024)
 * Modified references to ghosts/portraits to fit multigrid structure (19/11/2024)
 * Modify the mansion model (21/11/2024)
+* Modify rotation of the ghost prefab (21/11/2024)
 
 ### Deprecated